Update bootstrap scripts, idempotency with dconf setting
During bootstrap clear ansible directory before installing third party
modules. This ensure the directory is in a known good state.

When using the dconf module with a numeric value the module always
returns changed status when updated sucessfully; even if the vaule was
the same as before. Suppressing changed notice. If module fails it will
still generate an error.