Debian, use template to configure apt sources.list
This replaces the previous approach of editing the existing file to
enable contrib and non-free found in commit 2613d4e. Note that this does
not work with Buster and earlier as they have a different format for the
security repo. However Buster is only run on Chromebooks which do not
get contrib and non-free enabled.