Install Steam client on select workstations
For workstation role the variable steam_install can be used to install
the steam client on hosts. It has three options.

"false" does not install the steam client
"native" installs steam and i386 dependecies from repos, this is needed
for using Proton
"flatpak" installs steam from flathub, note that proton 5.13+ are not
supported on flatpak steam